Duplicate EPP::Server#initialize docs to EPP::Client.
The options hash which may be passed to EPP::Client#initialize is
passed through to the EPP::Server#initialize. In order to present
a better front for the primary interface in terms of documentation
the descriptions of these options has been duplicated to the docs
for the EPP::Client#initialize method.