Regression test: <domain:hostAttr> not properly supported.
According the RFC-5731 a domain's name servers can be specified in 2
ways: "Name servers are specified within a <domain:ns> element. This
element MUST contain one or more <domain:hostObj> elements or one or
more <domain:hostAttr> elements."

When <domain:hostAttr> is used the name and IP addresses should be
specified as follows:

  <domain:ns>
   <domain:hostAttr>
     <domain:hostName>ns1.example.net</domain:hostName>
     <domain:hostAddr
      ip="v4">192.0.2.2</domain:hostAddr>
     <domain:hostAddr
      ip="v6">1080:0:0:0:8:800:200C:417A</domain:hostAddr>
   </domain:hostAttr>
  </domain:ns>

But the code expects <domain:name> and <domain:addr> nodes instead
resulting in a NoMethodError.